move scope logic in its own file
