+++ openstack/datadog_checks/openstack/settings.py
@@ -1,0 +1,5 @@
+# (C) Datadog, Inc. 2010-2017
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+DEFAULT_KEYSTONE_API_VERSION = 'v3'
+DEFAULT_API_REQUEST_TIMEOUT = 10  # seconds
